django_server/requirements.txt
asgiref==3.8.1
bcrypt==4.1.2
cffi==1.16.0
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
easysnmp==0.2.6
et-xmlfile==1.1.0
Markdown==3.6
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.1
paramiko==3.4.0
pycparser==2.22
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
